improve mapbox text weight and italic handling
